+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## 1.5.0
 Expected: July 2025
 
+### New features
+
+* [NACM for Clixon Controller](https://github.com/clicon/clixon-controller/issues/189)
+
 ### API changes on existing protocol/config features
 
 * New `clixon-controller-config@2025-05-01.yang` revision
-  * Added `CONTROLLER_SERVICES_USER`
 
 ### Corrected Bugs